refactor: split ServiceForm into smaller files
