+++ packages/react-saasify/src/components/ServiceForm/ServiceEntry.js
@@ -1,0 +1,6 @@
+import React from 'react'
+
+// props is the "properties" object, where the values are React components
+export const ServiceEntry = ({ propKey, _type, onChange, ...props }) => (
+  <React.Fragment>{Object.values(props)}</React.Fragment>
+)
