+++ packages/react-saasify/src/components/ServiceForm/ServiceInputCheckbox.js
@@ -1,0 +1,22 @@
+import React from 'react'
+import { ServiceInputWrapper } from './ServiceInputWrapper'
+import { Checkbox } from 'lib/antd'
+
+export const ServiceInputCheckbox = ({
+  default: defaultValue,
+  type,
+  propKey,
+  onChange,
+  name
+}) => (
+  <ServiceInputWrapper>
+    <Checkbox
+      type={type}
+      checked={defaultValue}
+      onChange={onChange}
+      name={name}
+    >
+      {propKey}
+    </Checkbox>
+  </ServiceInputWrapper>
+)
